+++ src/Grid/GridConfig.php
@@ -9,7 +9,9 @@ use SebastianBergmann\CodeCoverage\Driver\Selector;
 class GridConfig
 {
     protected string $name = 'grid';
-    protected ?QueryBuilder $queryBuilder = null;
+    protected QueryBuilder $queryBuilder;
+    protected string $countFieldName;
+
     protected ?PaginatorConfig $paginatorConfig = null;
 
     /** @var Field[] */
@@ -18,7 +20,11 @@ class GridConfig
     /** @var Selector[] */
     protected array $selectorList = [];
 
-    protected ?string $countFieldName = null;
+    public function __construct(QueryBuilder $queryBuilder, string $countFieldName)
+    {
+        $this->queryBuilder = $queryBuilder;
+        $this->countFieldName = $countFieldName;
+    }
 
     /**
      * @param Field|string $field
@@ -57,9 +63,6 @@ class GridConfig
         return $this->name;
     }
 
-    /**
-     * @return GridConfig Fluent interface
-     */
     public function setQueryBuilder(QueryBuilder $queryBuilder): self
     {
         $this->queryBuilder = $queryBuilder;
@@ -67,7 +70,7 @@ class GridConfig
         return $this;
     }
 
-    public function getQueryBuilder(): ?\Doctrine\ORM\QueryBuilder
+    public function getQueryBuilder(): QueryBuilder
     {
         return $this->queryBuilder;
     }
@@ -105,7 +108,7 @@ class GridConfig
     /**
      * Returns the field corresponding to the name
      */
-    public function getFieldByName(string $name): ?Field
+    public function getFieldByName(string $name): Field
     {
         foreach ($this->fieldList as $field) {
             if ($field->getFieldName() === $name) {
@@ -113,7 +116,7 @@ class GridConfig
             }
         }
 
-        return null;
+        throw new \InvalidArgumentException(sprintf('There is no defined field by name "%s".', $name));
     }
 
     /**
@@ -131,16 +134,6 @@ class GridConfig
         }
 
         return $matchingFieldList;
-    }
-
-    /**
-     * @return GridConfig Fluent interface
-     */
-    public function setCountFieldName(string $countFieldName): self
-    {
-        $this->countFieldName = $countFieldName;
-
-        return $this;
     }
 
     public function getCountFieldName(): string