+++ CHANGELOG.md
@@ -2,6 +2,7 @@ Changelog
 =========
 
 ## Unreleased
+- Filter fields in case-insensitive mode
 
 ## 4.0.1 - 2022-03-25
 - just tag for a correct commit