+++ CHANGELOG.md
@@ -3,6 +3,8 @@ Changelog
 
 ## Unreleased
 - Filter fields in case-insensitive mode
+- [**BC**] Fix nullable types
+- [**BC**] Require mandatory parameters in constructor
 
 ## 4.0.1 - 2022-03-25
 - just tag for a correct commit